move generators to new autopas-tools target and directory
