+++ examples/md-flexible/CMakeLists.txt
@@ -3,7 +3,6 @@ file(
     MDFlex_SRC
     "*.cpp"
     "*.h"
-    "${PROJECT_SOURCE_DIR}/tests/testAutopas/testingHelpers/RandomGenerator.cpp"
 )
 
 # remove all files from the md-flexible/tests dir
@@ -20,7 +19,7 @@ target_include_directories(md-flexible PUBLIC ${CMAKE_CURRENT_SOURCE_DIR})
 
 include(autopas_yaml-cpp)
 
-target_link_libraries(md-flexible PUBLIC autopas yaml-cpp)
+target_link_libraries(md-flexible PUBLIC autopas autopas-tools yaml-cpp)
 
 # --- copy script files to build dir ---
 file(
