+++ examples/md-flexible/tests/TimeDiscretizationTest.cpp
@@ -7,8 +7,8 @@
 #include "TimeDiscretizationTest.h"
 
 #include "TimeDiscretization.h"
+#include "autopas-tools/generators/GridGenerator.h"
 #include "autopas/utils/ArrayMath.h"
-#include "testingHelpers/GridGenerator.h"
 
 void TimeDiscretizationTest::fillWithParticlesAndInit(autopas::AutoPas<Molecule, FMCell> &autopas) {
   autopas.setBoxMin({0., 0., 0.});
