+++ src/autopas/containers/verletListsCellBased/verletLists/neighborLists/asBuild/VerletNeighborListAsBuild.h
@@ -6,9 +6,9 @@
 
 #pragma once
 
-#include "../VerletNeighborListInterface.h"
 #include "AsBuildPairGeneratorFunctor.h"
 #include "C08TraversalColorChangeNotify.h"
+#include "autopas/containers/verletListsCellBased/verletLists/neighborLists/VerletNeighborListInterface.h"
 #include "autopas/utils/WrapOpenMP.h"
 
 namespace autopas {
