+++ tests/testAutopas/tests/containers/LinkedCellsVersusVerletClusterListsTest.h
@@ -11,9 +11,9 @@
 #include <cstdlib>
 
 #include "AutoPasTestBase.h"
+#include "autopas-tools/generators/RandomGenerator.h"
 #include "autopas/autopasIncludes.h"
 #include "autopas/molecularDynamics/ParticlePropertiesLibrary.h"
-#include "testingHelpers/RandomGenerator.h"
 #include "testingHelpers/commonTypedefs.h"
 
 class LinkedCellsVersusVerletClusterListsTest : public AutoPasTestBase {
