+++ tests/testAutopas/tests/containers/LinkedCellsVersusVerletListsTest.h
@@ -12,9 +12,9 @@
 #include <memory>
 
 #include "AutoPasTestBase.h"
+#include "autopas-tools/generators/RandomGenerator.h"
 #include "autopas/autopasIncludes.h"
 #include "autopas/molecularDynamics/ParticlePropertiesLibrary.h"
-#include "testingHelpers/RandomGenerator.h"
 #include "testingHelpers/commonTypedefs.h"
 
 class LinkedCellsVersusVerletListsTest : public AutoPasTestBase {
