+++ tests/testAutopas/tests/containers/directSum/DirectSumTraversalTest.cpp
@@ -6,8 +6,8 @@
 
 #include "DirectSumTraversalTest.h"
 
+#include "autopas-tools/generators/RandomGenerator.h"
 #include "autopas/containers/directSum/DirectSumTraversal.h"
-#include "testingHelpers/RandomGenerator.h"
 
 using ::testing::_;
 using ::testing::AtLeast;
