+++ tests/testAutopas/tests/containers/linkedCells/traversals/C01TraversalTest.h
@@ -9,9 +9,9 @@
 #include <gtest/gtest.h>
 
 #include "AutoPasTestBase.h"
+#include "autopas-tools/generators/GridGenerator.h"
 #include "autopas/AutoPas.h"
 #include "mocks/MockFunctor.h"
-#include "testingHelpers/GridGenerator.h"
 #include "testingHelpers/commonTypedefs.h"
 
 class C01TraversalTest : public AutoPasTestBase {
