+++ tests/testAutopas/tests/containers/linkedCells/traversals/C08TraversalTest.h
@@ -9,9 +9,9 @@
 #include <gtest/gtest.h>
 
 #include "AutoPasTestBase.h"
+#include "autopas-tools/generators/GridGenerator.h"
 #include "autopas/AutoPas.h"
 #include "mocks/MockFunctor.h"
-#include "testingHelpers/GridGenerator.h"
 #include "testingHelpers/commonTypedefs.h"
 
 class C08TraversalTest : public AutoPasTestBase {
