+++ tests/testAutopas/tests/containers/linkedCells/traversals/SlicedTraversalTest.cpp
@@ -6,7 +6,7 @@
 
 #include "SlicedTraversalTest.h"
 
-#include "testingHelpers/GridGenerator.h"
+#include "autopas-tools/generators/GridGenerator.h"
 #include "testingHelpers/NumThreadGuard.h"
 
 using ::testing::_;
