+++ tests/testAutopas/tests/sph/SPHTest.h
@@ -11,9 +11,9 @@
 #include <tuple>
 
 #include "AutoPasTestBase.h"
+#include "autopas-tools/generators/RandomGenerator.h"
 #include "autopas/autopasIncludes.h"
 #include "autopas/sph/autopassph.h"
-#include "testingHelpers/RandomGenerator.h"
 
 enum SPHFunctorType { density, hydro };
